requirements.txt
dill==0.2.8.2
matplotlib==3.0.0
notebook==5.7.0
numpy==1.15.2
pandas==0.23.4
python==3.6.6
pytorch==0.4.1
torchtext==0.3.1
tqdm==4.26.0
urllib3==1.23
